refactor(auth): move functions from `utils/auth` to `models/user`
